Switch from one pool of passwords to per-namespace pools
We want to authenticate EPIC API users based on credentials in their
account namespaces, so we need to be more specific about which
credentials are used to authenticate a given request. The Htpasswd
method hoovered up all of the passwords in all of the secrets and put
them into one pool. We need to keep a credential pool for each
namespace so we can authenticate requests in that namespace with
credentials in that namespace.